Extract migrations to F2 module
