+++ features/F2/__init__.py
@@ -1,5 +1,5 @@
 """Feature F2 package."""
 
-from . import duplicate_finder, metadata_store, path_links
+from . import duplicate_finder, metadata_store, path_links, migrations
 
-__all__ = ["metadata_store", "path_links", "duplicate_finder"]
+__all__ = ["metadata_store", "path_links", "duplicate_finder", "migrations"]
